requirements.txt
Django==2.2.2
djangorestframework==3.11.2
django-jet==1.0.8
django-rest-auth==0.9.3
django-coreapi
django-smtp-ssl
django-filter==2.2.0
django-url-filter
gunicorn==19.9.0
drf_yasg==1.17.0
pillow==8.2.0
face-recognition==1.3.0
gunicorn==19.9.0
psycopg2-binary==2.8.6
opencv-contrib-python==4.5.3.56